Add iovec HAL API and adapt transports
Introduce a scatter-gather HAL API and refactor transports and ports to use it. Key changes:

- Add tinypan_iovec_t and hal_bt_l2cap_send_iovec() to include/tinypan_hal.h.
- Implement iovec-based send in hal/mock, esp32, and zephyr HALs; esp32/zephyr also add hal_get_local_bd_addr and improve thread-safe RX/event bridging.
- Refactor BNEP transport to build an iovec from lwIP pbuf chains and avoid draining the TX queue from the lwIP thread; fix payload-slice logic and offset type.
- Rework SLIP transport to a streaming/chunked encoder (smaller chunk buffer, incremental state machine) to reduce peak RAM and handle busy sends.
- Improve ESP32 HAL: aligned bounce buffer handling, event/rx queues, safe copy from BT callbacks, and hal_bt_poll for processing queued events/data.
- Update README and configuration comments to reflect architecture, memory model, and uncompressed BNEP headers.
- Minor test comment wording change.

These changes enable zero-copy/scatter-gather transmission while improving thread-safety and reducing transient memory usage for constrained MCUs.

+++ src/tinypan_bnep.c
@@ -335,7 +335,7 @@ int bnep_parse_ethernet_frame(const uint8_t* data, uint16_t len,
         return -1;
     }
     
-    uint16_t ext_offset = header_len;
+    uint32_t ext_offset = header_len;
     while (has_ext) {
         if (ext_offset + 2 > len) {
             TINYPAN_LOG_WARN("Packet too short for BNEP extension header");
